backend 2/requirements.txt
asgiref==3.11.1
attrs==25.4.0
black==26.1.0
certifi==2026.1.4
cffi==2.0.0
cfgv==3.5.0
charset-normalizer==3.4.4
click==8.3.1
cloudinary==1.44.1
coverage==7.13.4
cryptography==46.0.5
distlib==0.4.0
dj-database-url==3.1.0
dj-rest-auth==7.0.2
Django==5.2.11
django-allauth==65.14.3
django-anymail==14.0
django-cloudinary-storage==0.3.0
django-cors-headers==4.9.0
django-debug-toolbar==6.2.0
django-extensions==4.1
django-filter==25.2
django-rich==2.2.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
factory_boy==3.3.3
Faker==40.4.0
filelock==3.21.2
identify==2.6.16
idna==3.11
inflection==0.5.1
iniconfig==2.3.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
model-bakery==1.23.3
mypy_extensions==1.1.0
nodeenv==1.10.0
oauthlib==3.3.1
packaging==26.0
pathspec==1.0.4
pillow==12.1.1
platformdirs==4.7.1
pluggy==1.6.0
pre_commit==4.5.1
psycopg2-binary==2.9.11
pycparser==3.0
Pygments==2.19.2
PyJWT==2.11.0
pytest==9.0.2
pytest-django==4.12.0
python-decouple==3.8
pytokens==0.4.1
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rich==14.3.3
rpds-py==0.30.0
ruff==0.15.1
six==1.17.0
sqlparse==0.5.5
uritemplate==4.2.0
urllib3==2.6.3
virtualenv==20.36.1
Werkzeug==3.1.6
whitenoise==6.11.0
